removed PMD warnings and trailing whitespace

+++ zmpp-core/src/main/java/org/zmpp/base/DefaultStoryFileHeader.java
@@ -22,7 +22,7 @@ import static org.zmpp.base.MemoryUtil.*;
 
 /**
  * This is the default implementation of the StoryFileHeader interface.
- * 
+ *
  * @author Wei-ju Wu
  * @version 1.5
  */
@@ -30,7 +30,7 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
 
   /** The memory map. */
   private Memory memory;
-  
+
   /**
    * Constructor.
    * @param memory a Memory object
@@ -41,10 +41,10 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
 
   /** {@inheritDoc} */
   public int getVersion() { return memory.readUnsigned8(0x00); }
-  
+
   /** {@inheritDoc} */
   public String getSerialNumber() { return extractAscii(0x12, 6); }
-  
+
   /** {@inheritDoc} */
   public int getFileLength() {
     // depending on the story file version we have to multiply the
@@ -59,7 +59,7 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
     }
     return fileLength;
   }
-  
+
   /**
    * {@inheritDoc}
    */
@@ -70,7 +70,7 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
       memory.writeUnsigned8(0x1f, (char) version);
     }
   }
-  
+
   /**
    * {@inheritDoc}
    */
@@ -81,7 +81,7 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
       memory.writeUnsigned8(0x26, (char) units);
     }
   }
-  
+
   /**
    * {@inheritDoc}
    */
@@ -92,7 +92,7 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
       memory.writeUnsigned8(0x27, (char) units);
     }
   }
-  
+
   /** {@inheritDoc} */
   public void setMouseCoordinates(final int x, final int y) {
     // check the extension table
@@ -107,12 +107,12 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
       }
     }
   }
-  
+
   /** {@inheritDoc} */
   public char getCustomAccentTable() {
     // check the extension table
     char result = 0;
-    final int extTable = memory.readUnsigned16(0x36);    
+    final int extTable = memory.readUnsigned16(0x36);
     if (extTable > 0) {
       final int numwords = memory.readUnsigned16(extTable);
       if (numwords >= 3) {
@@ -121,11 +121,11 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
     }
     return result;
   }
-  
+
   // ***********************************************************************
   // ****** Attributes
   // **********************************
-  
+
   /**
    * {@inheritDoc}
    */
@@ -165,7 +165,7 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
       break;
     }
   }
-  
+
   /**
    * {@inheritDoc}
    */
@@ -191,11 +191,12 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
   public String toString() {
     StringBuilder builder = new StringBuilder();
     for (int i = 0; i < 55; i++) {
-      builder.append(String.format("Addr: %02x Byte: %02x\n", i, memory.readUnsigned8(i)));
+      builder.append(String.format("Addr: %02x Byte: %02x\n", i,
+      														 (int) memory.readUnsigned8(i)));
     }
     return builder.toString();
   }
-  
+
   // ************************************************************************
   // ****** Private section
   // *******************************
@@ -203,7 +204,7 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
   /**
    * Extract an ASCII string of the specified length starting at the specified
    * address.
-   * 
+   *
    * @param address the start address
    * @param length the length of the ASCII string
    * @return the ASCII string at the specified position
@@ -211,71 +212,71 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
   private String extractAscii(final int address, final int length) {
     final StringBuilder builder = new StringBuilder();
     for (int i = address; i < address + length; i++) {
-      
+
       builder.append((char) memory.readUnsigned8(i));
     }
     return builder.toString();
   }
-    
+
   private void setTranscripting(final boolean flag) {
     char flags = memory.readUnsigned16(0x10);
     flags = (char) (flag ? (flags | 1) : (flags & 0xfe));
     memory.writeUnsigned16(0x10, (char) flags);
   }
-  
+
   private boolean isTranscriptingOn() {
     return (memory.readUnsigned16(0x10) & 1) > 0;
   }
-  
+
   private boolean forceFixedFont() {
     return (memory.readUnsigned16(0x10) & 2) > 0;
   }
-  
+
   private void setForceFixedFont(final boolean flag) {
     char flags = memory.readUnsigned16(0x10);
     flags = (char) (flag ? (flags | 2) : (flags & 0xfd));
     memory.writeUnsigned16(0x10, (char) flags);
   }
-  
+
   private void setTimedInputAvailable(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
     flags = flag ? (flags | 128) : (flags & 0x7f);
     memory.writeUnsigned8(0x01, (char) flags);
   }
-  
+
   private boolean isScoreGame() {
     return (memory.readUnsigned8(0x01) & 2) == 0;
   }
-  
+
   private void setFixedFontAvailable(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
     flags = flag ? (flags | 16) : (flags & 0xef);
     memory.writeUnsigned8(0x01, (char) flags);
   }
-  
+
   private void setBoldFaceAvailable(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
     flags = flag ? (flags | 4) : (flags & 0xfb);
     memory.writeUnsigned8(0x01, (char) flags);
-  }  
+  }
 
   private void setItalicAvailable(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
     flags = flag ? (flags | 8) : (flags & 0xf7);
     memory.writeUnsigned8(0x01, (char) flags);
   }
-  
+
   private void setScreenSplittingAvailable(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
     flags = flag ? (flags | 32) : (flags & 0xdf);
     memory.writeUnsigned8(0x01, (char) flags);
   }
-  
+
   private void setStatusLineAvailable(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
     flags = flag ? (flags | 16) : (flags & 0xef);
     memory.writeUnsigned8(0x01, (char) flags);
-  }  
+  }
 
   private void setDefaultFontIsVariablePitch(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
@@ -286,13 +287,13 @@ public class DefaultStoryFileHeader implements StoryFileHeader {
   private boolean defaultFontIsVariablePitch() {
     return (memory.readUnsigned8(0x01) & 64) > 0;
   }
-  
+
   private void setSupportsColours(final boolean flag) {
     int flags = memory.readUnsigned8(0x01);
     flags = flag ? (flags | 1) : (flags & 0xfe);
     memory.writeUnsigned8(0x01, (char) flags);
   }
-  
+
   private boolean useMouse() {
     return (memory.readUnsigned8(0x10) & 32) > 0;
   }
